post-restore script for restores with DCs and defaultVolumesToRestic
When restoring a backup with defaultVolumesToRestic=true and
DeploymentConfigs in the backup, this script must be run after the
completion of the restore to clean up the restic restore actions.

The openshift plugins will disconnect DC pods from the ReplicationController
when defaultVolumesToRestic is true and set DC replicas to 0.
This script will remove the restore pods and reset DC replicas so that
new pods will be created that are properly reconnected to the DC.