add a simple golang version check to warn developers
For new developers it may not be clear why some of the make commands
are not working properly.  For example [1] changed the minimum
required golang to successfully use `make deploy-olm`. It was
also not 100% clear what golang version would work across platforms.

It may be helpful to specify the required version and expose it
via the makefile.

[1]https://github.com/openshift/oadp-operator/pull/809