Disable cluster monitoring for Velero and remove ServiceMonitor from the Velero deployment. (#1092)
First change to remove cluster monitoring of OADP that will be replaced by the user workload
monitoring (UWM).

Related Issues:
  https://github.com/openshift/oadp-operator/issues/769
  https://issues.redhat.com/browse/OADP-1887
  https://issues.redhat.com/browse/OADP-661

The openshift-adp-velero-metrics-svc is left to easy process of enabling UWM.

Once UWM is enabled, it will require setting up ServiceMonitor and configuring alerts
or dashboards that are crucial for a particular use-case.

Enablement of user workload monitoring with additional documentation is not part
of this PR to easilly re-add the cluster monitoring in the future by reverting this change.

Signed-off-by: Michal Pryc <mpryc@redhat.com>
Co-authored-by: Michal Pryc <mpryc@redhat.com>