New infrastructure annotation format (#1155)
Cherry-Pick: https://github.com/openshift/oadp-operator/pull/1155

* New infrastructure annotation format

https://docs.engineering.redhat.com/display/CFC/Best_Practices#Best_Practices-(New)RequiredInfrastructureAnnotations

The release pipelines will start to block in case of absence of the
annotations in the namespace features.operators.openshift.io in Q1 2024.
The old annotations weren't enforced in the pipeline because of their
format providing no way to differentiate between deliberate absence of a
value from human obliviousness. Support for the new annotations will be
backported all the way to OpenShift 4.10.

Latest by Q1 2024, you MUST provide either the string value (not
boolean) "true" or "false" for each of the following list of
annotations:

    features.operators.openshift.io/disconnected: "true"
    features.operators.openshift.io/fips-compliant: "false"
    features.operators.openshift.io/proxy-aware: "false"
    features.operators.openshift.io/cnf: "false"
    features.operators.openshift.io/cni: "false"
    features.operators.openshift.io/csi: "false"
    features.operators.openshift.io/tls-profiles: "false"
    features.operators.openshift.io/token-auth-aws: "false"
    features.operators.openshift.io/token-auth-azure: "false"
    features.operators.openshift.io/token-auth-gcp: "false"

You MUST provide a value for all of these fields.

---------

Co-authored-by: Tiger Kaovilai <tkaovila@redhat.com>
Signed-off-by: Rayford Johnson <rayfordj@users.noreply.github.com>