OADP-1737: [oadp-1.2] Fix defaultl credentials logic
A logic which was looping over BSL and VSL then setting requirement
for default credentials was broken for the SnapshotLocations, where
it was missing check if the provider was already set in the map.

We should only set the need for default credentials for the given
provider if it was not already found in the map.

This fixes OADP-1737 in which the BSL was going to an unavailable
phase when VSL credentials were specified as there were no credentials
mounted to the velero pod nor the AWS_SHARED_CREDENTIALS_FILE env
specified for that pod.

Signed-off-by: Michal Pryc <mpryc@redhat.com>
Co-authored-by: Scott Seago <sseago@redhat.com>