OADP-1.2 Cherry-pick of OADP-659: Remove HTTP/HTTPS port numbers from AWS S3 URLs.  (#1180)
* OADP-659 Remove HTTP/HTTPS port numbers from AWS S3 URLs. (#1169)

* OADP-659 Remove port numbers if HTTP or HTTPS.

Avoids a signature mismatch from S3-compatible services.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* OADP-659 Move StripDefaultPorts to pkg/common.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* OADP-659 Fix port number in comment.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

---------

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* OADP-659 Avoid overwriting good hostname. (#1177)

SanitizeHostForHeader puts the stripped host name into the Host field of
the HTTP request, but only if it actually needed to strip the port. If
there is already no port, that Host field can be blank, and there is no
reason to overwrite the URL's host field.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

---------

Signed-off-by: Matthew Arnold <marnold@redhat.com>