Remove VSB and VSR from OADP 1.3+
Since VSB and VSR is not needed in OADP 1.3+ removing it from the CSV.

Updated API_ref and script to not show errors. E.g. when the ocp is missing
those APIs, however left the functions in the datamover_resources.sh as one
may want to run it against earlier OADP versions.