[oadp-1.3] Add flake attempts to the flaky tests and include known flakes pattern detection (#1297)
* Add livenessProbe and startupProbe to the sample-applications

This change adds the liveness and startup probe to the mongo
and mysql applications within the sample applications that are used
during CI runs.

The livenessProbe checks if the relevant port is open within the container
run, where the startupProbe checks if it's possible to actually
connect to the DB endpoint from the container in which database
is running.

It also checks for the correct condition of the running pod and not only
for the state in case probe fails.

Increased timeout in the AreApplicationPodsRunning which is ran via
Eventually to 19minuts which should be more then enough time for even
complex scenarios.

Removed sleep as it should not be needed anymore as we are checking for
the startup probe now.

Signed-off-by: Michal Pryc <mpryc@redhat.com>

* Add flake attempts to the flaky tests and include known flakes pattern detection

This adds attempts to the flaky tests, which are caused by known
issues.

The known flake pattern detection allows us to specify which flakes
are the ones on which we will retry.

Signed-off-by: Michal Pryc <mpryc@redhat.com>

---------

Signed-off-by: Michal Pryc <mpryc@redhat.com>
Co-authored-by: Michal Pryc <mpryc@redhat.com>