OADP-2982: Test VMs from YAML templates (#1359)
* Create VM with YAML template.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Install VM disk from YAML too.

Rework virtualization test to pick up DataVolume YAML from the
sample-applications/virtual-machines directory, using the namespace as a
subdirectory and YAML files with "-disk.yaml" and "-vm.yaml" suffixes.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Add more comments about DataVolume workflow.

Hopefully a better description about what the test is doing with
DataVolumes/VMs and why.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Remove actual YAML file reference from comments.

Just in case cirros-test needs to be removed.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Add a DataVolume behavioral note to test YAML.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

---------

Signed-off-by: Matthew Arnold <marnold@redhat.com>