Avoid trying to delete non-existent storage class.
If the test did not need to create an immediate-mode storage class
(because the default was already set to immediate), make sure it does
not try to delete anything after the test is done. Previously this would
cause an error after all the tests were run, even if they were all
successful.

Signed-off-by: Matthew Arnold <marnold@redhat.com>