Virtualization E2E tests on OpenStack (#1511)
* Avoid deleting previously installed CirrOS image.

Only clean it up if the test downloaded it in the first place.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Default OpenStack provider to no KVM emulation.

Allow KVM emulation to be configured, and add an OpenStack provider that
copies AWS except for leaving emulation turned off.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Avoid trying to delete non-existent storage class.

If the test did not need to create an immediate-mode storage class
(because the default was already set to immediate), make sure it does
not try to delete anything after the test is done. Previously this would
cause an error after all the tests were run, even if they were all
successful.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Create WFFC storage class if default is Immediate.

Otherwise the previously-default WFFC tests will all run with Immediate,
and make the Immediate-specific tests redundant.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Add test scripts for OpenStack.

Can't just use AWS for everything after all.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

---------

Signed-off-by: Matthew Arnold <marnold@redhat.com>