Make common function for updating BSL spec (#1631)
* Make common function for updating BSL spec

With the self service project (Non Admin) it is required
to update BSL spec based on the same conditions as OADP BSL
spec would be updated. It makes sense to move out the logic
from the bsl specific function to a common publicly available
from which NABSL controller can also benefit.

Signed-off-by: Michal Pryc <mpryc@redhat.com>

* Add nolint unparam to reworked UpdateBackupStorageLocation function

Rebase and added nolint to ensure oadp enabled linting is working.

Signed-off-by: Michal Pryc <mpryc@redhat.com>

---------

Signed-off-by: Michal Pryc <mpryc@redhat.com>