Add NonAdminBSL Approval CRD with option to enable/disable it (#1633)
* Add NonAdminBSL Approval CRD with option to enable/disable admin approval for NABSL

Adds CRD that allows admin user to decide if the NABSL needs to be approved prior
to creating BSL object.

Part-of: #204

Signed-off-by: Michal Pryc <mpryc@redhat.com>

* Update manifests after review

Updated manifests to include default description for the new
RequireAdminApprovalForBSL option and corrected description for this
field.

Signed-off-by: Michal Pryc <mpryc@redhat.com>

* Rename RequireAdminApprovalForBSL to RequireApprovalForBSL

Makes the DPA setting shorter.

Signed-off-by: Michal Pryc <mpryc@redhat.com>

* Rename status field for the nonAdminBSL spec.

Rename Status to include nonAdminBackupStorageLocation
instead of veleroBackupStorageLocationRequest.

Signed-off-by: Michal Pryc <mpryc@redhat.com>

---------

Signed-off-by: Michal Pryc <mpryc@redhat.com>