Update podConfig and loadAffinity to ensure proper scheduling
In this update we ensure the nodeSelector within podConfig is scheduled
in less restricted way then the loadAffinity, ensuring the DataMover
pods will get scheduled on the nodes where node-agents are running.

This corner case was when user used both settings and used different
labels between podConfig and loadAffinity.

The loadAffinity must be a subset of what's defined in the podConfig.

Signed-off-by: Michal Pryc <mpryc@redhat.com>