Node Selector Configuration for Data Protection Components - NodeAgent and DataMover (#1683)
* Node Selector Configuration for Data Protection Components

This proposal introduces a standardized approach for configuring
node selectors in the DataProtectionApplication (DPA)
custom resource for various use cases that are common
across different Velero settings.

Fixes: OADP-5832

Signed-off-by: Michal Pryc <mpryc@redhat.com>

* Update design to not cover Velero pod in the current implementation

The Velero pod uses node selector and should be sufficient for most
cases, because there is only one pod scheduled and as such user have
1-1 mapping with the node label on which the pod is scheduled.

Signed-off-by: Michal Pryc <mpryc@redhat.com>

* Update podConfig and loadAffinity to ensure proper scheduling

In this update we ensure the nodeSelector within podConfig is scheduled
in less restricted way then the loadAffinity, ensuring the DataMover
pods will get scheduled on the nodes where node-agents are running.

This corner case was when user used both settings and used different
labels between podConfig and loadAffinity.

The loadAffinity must be a subset of what's defined in the podConfig.

Signed-off-by: Michal Pryc <mpryc@redhat.com>

---------

Signed-off-by: Michal Pryc <mpryc@redhat.com>